MOVE all PIP template parts to skeleton-implementations
