+++ pwa/src/skeleton-implementations/pip/layout/AuthenticatedLayout.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import * as styles from "./Layout.module.css";
-import { AuthenticatedFooterTemplate } from "../../../templates/templateParts/footer/FooterTemplate";
-import { AuthenticatedHeaderTemplate } from "../../../templates/templateParts/header/HeaderTemplate";
+import { AuthenticatedFooterTemplate } from "../templates/templateParts/footer/FooterTemplate";
+import { AuthenticatedHeaderTemplate } from "../templates/templateParts/header/HeaderTemplate";
 import { PrivateRoute } from "@conduction/components";
 import { isLoggedIn } from "../../../services/auth";
