+++ pwa/src/skeleton-implementations/pip/pages/index.tsx
@@ -1,8 +1,8 @@
 import * as React from "react";
 import { isLoggedIn } from "../../../services/auth";
-import { DashboardTemplate } from "../../../templates/dashboard/DashboardTemplate";
-import { LandingTemplate } from "../../../templates/landing/LandingTemplate";
+import { DashboardTemplate } from "../templates/dashboard/DashboardTemplate";
 import { HomeTemplate } from "../../../templates/templateParts/home/HomeTemplate";
+import { LandingTemplate } from "../templates/landing/LandingTemplate";
 
 const IndexPage: React.FC = () => {
   if (isLoggedIn()) return <AuthenticatedIndex />;
