+++ pwa/src/skeleton-implementations/pip/templates/templateParts/footer/FooterTemplate.tsx
