+++ pwa/src/skeleton-implementations/pip/templates/templateParts/home/HomeTemplate.tsx
